+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 ------------------
 
 - #2677 Fix permission for Uncertainty field
+- #2675 Fix services are not deselected on template removal in sample add form
+- #2673 Trigger recalculation of dependants if uncertainty changes
+- #2674 Fix partitions not displayed correctly in batch samples listing
 - #2672 Fix rejected sample analyses are re-added on profile removal
 - #2670 Flush calculated result if dependency is flushed
 - #2667 Specifications support for multi-result analyses